agent/pyproject.toml
[project]
name = "social-wingman-agent"
version = "0.1.0"
description = "Practice Partner agent using LiveKit and OpenAI Realtime"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[openai]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv~=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
